requirements.txt
python-telegram-bot
openai-whisper
anthropic
ffmpeg-python 